BroadleafCommerce/QA#1944 - Fix an issue where properties are not registered in all dependent properties prior to instantiation
In this case, this was caused by the order property not being set in the ValidateAndConfirmPaymentsActivity until _after_ the BaseProcessor (specifically the blCheckoutWorkflow) was already instantiated and its lifecycle methods executed. This resulted in the ValidateAndConfirmPaymentsActivity to have the wrong order in the final list of blCheckoutActivities.

Since we are sorting the list of activities in afterPropertiesSet() and since there is no 'real' dependency beans on the activities already being created, sorting of the activities list was moved to after spring's ApplicationContext completes refreshing which means all beans will have been instantiated and set