Return the amount in the currency of the transactions and not in the currency of the order
In the event that the Order currency is different than the PaymentTransaction(s) then what would previously happen is we would initialize "amount" with the Order's currency, then add the PaymentTransactions whose amounts would get converted to the Order's currency. What should happen is the PaymentTransaction amount should be returned in the currency that they were created with. We still want to make sure that the returned amount is non-null to be consistent with before however we will initialize the returned amount with the first PaymentTransaction instead of the Order's currency.

In the event that there's two or more PaymentTransactions that are of the requested type but their amounts are in two different currencies then the implementor will simply have to convert the amount to the correct currency in an override since it's non-deterministic.